Initialize fluid hardness to match vanilla fluids
This initialises BoP fluids to have a hardness of 100.0F which matches
vanilla water and lava. Calling #setHardness subsequently sets the
blockResistance (blockHardness * 5.0F) which handles resistence to
explosions.

An exception was made for quicksand, which initializes to the same
hardness as minecraft:sand (0.5F) so that explosions won't leave piles
of quicksand floating around.

Signed-off-by: srs-bsns <forfrdm@gmail.com>