Large API changes
Large unfinished changes to inner workings of Biome Registration.

BOPBiomes now use RegistryObject's rather than Optional's, as they self populate without need of reflection and provide things such as the ResourceLocation along with having same features as Optional's.

BiomeRegistry mostly no longer requires the use of DeferredRegistration type, and RegistrationData types now have a function which performs the work of the RegistrationType lambda's which now just invoke the RegistrationData's register() function.

RegistrationData types all use their register() function to provide a config entry if one isn't present, or to use the config entry if it is to perform the work that the RegistrationType lambda's previously were providing, this means the config functions are no longer required as register() does that work.

TO-DO: remove all usage of RegistrationType's and DeferredRegistration types, change the name of the defer functions, and simply create the new RegistrationType in that function and invoke it's register() function when called. Also figure out the best way to utilize Forge's DeferredRegister class and delay building of BOP climate mappings until after registry injection is finalized to justify using Suppliers, or use pure Biome types if too much.