Provide compatibility enhancements
Simplifies complexity by removing the need to use reflection to get the values of the vanilla entries, along with removing the need to create deep copies of each climates land and island biomes to check whether a biome entry already exists in a given climate.

Checks the BiomeDictionary to see if the biome is a NETHER biome, as the NETHER climate provides no BiomeType making it impossible to scrape modded nether biomes using a BiomeType comparison due to the inner workings of Forge's BiomeManager class itself.